+++ packages/db/schema/index.ts
@@ -16,4 +16,5 @@ export * from './payments';
 export * from './subscriptions';
 export * from './reconciliation';
 export * from './audit';
+export * from './cms';
 export * from './views';